requirements.txt
loguru
torch>=1.10.0
onnx>=1.10.2
onnxruntime>=1.9.0
onnx-simplifier>=0.3.6
torchinfo